2019/requirements.txt
#core packages
networkx==2.4
numpy==1.17.4
python-dotenv==0.10.3

#exploration packages
jupyter
matplotlib
pandas
scipy
seaborn

#local package
-e .